Refactor and reorganize test utilities for internal use
