+++ doma-processor/src/test/java/org/seasar/doma/internal/apt/CompilerSupport.java
@@ -15,8 +15,6 @@
  */
 package org.seasar.doma.internal.apt;
 
-import static org.junit.jupiter.api.Assertions.fail;
-
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
@@ -24,12 +22,9 @@ import java.io.InputStreamReader;
 import java.io.UncheckedIOException;
 import java.net.URL;
 import java.nio.file.Path;
-import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import javax.annotation.processing.Processor;
-import javax.tools.Diagnostic;
-import javax.tools.JavaFileObject;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.jupiter.api.io.TempDir;
@@ -37,43 +32,39 @@ import org.seasar.doma.message.Message;
 
 public abstract class CompilerSupport {
 
-  @RegisterExtension final CompilerExtension compiler = new CompilerExtension();
+  private @RegisterExtension final AptinaTestCase delegate = new AptinaTestCase();
 
-  @TempDir Path sourceOutput;
+  private @TempDir Path sourceOutput;
 
-  @TempDir Path classOutput;
+  private @TempDir Path classOutput;
 
   @BeforeEach
   final void setupTempDirs() {
-    compiler.setSourceOutput(sourceOutput);
-    compiler.setClassOutput(classOutput);
+    delegate.setSourceOutput(sourceOutput);
+    delegate.setClassOutput(classOutput);
   }
 
   protected void enableCompilationAssertion() {
-    compiler.enableCompilationAssertion();
+    delegate.enableCompilationAssertion();
   }
 
   protected void disableCompilationAssertion() {
-    compiler.disableCompilationAssertion();
-  }
-
-  protected void addSourcePath(final String... sourcePaths) {
-    compiler.addSourcePath(sourcePaths);
+    delegate.disableCompilationAssertion();
   }
 
   protected void addOption(final String... options) {
     if (options.length == 0) {
       return;
     }
-    compiler.addOption(options);
+    delegate.addOption(options);
   }
 
   protected void addProcessor(final Processor... processors) {
-    compiler.addProcessor(processors);
+    delegate.addProcessor(processors);
   }
 
   protected void addCompilationUnit(final Class<?>... classes) {
-    compiler.addCompilationUnit(classes);
+    delegate.addCompilationUnit(classes);
   }
 
   protected void addResourceFileCompilationUnit(final String fqn) {
@@ -82,7 +73,7 @@ public abstract class CompilerSupport {
       Objects.requireNonNull(in);
       try (BufferedReader reader = new BufferedReader(new InputStreamReader(in))) {
         String source = reader.lines().collect(Collectors.joining("\n"));
-        compiler.addCompilationUnit(fqn, source);
+        delegate.addCompilationUnit(fqn, source);
       }
     } catch (IOException e) {
       throw new UncheckedIOException(e);
@@ -90,36 +81,23 @@ public abstract class CompilerSupport {
   }
 
   protected void compile() throws IOException {
-    compiler.compile();
+    delegate.compile();
   }
 
   protected Boolean getCompiledResult() throws IllegalStateException {
-    return compiler.getCompiledResult();
-  }
-
-  protected List<Diagnostic<? extends JavaFileObject>> getDiagnostics() {
-    return compiler.getDiagnostics();
+    return delegate.getCompiledResult();
   }
 
   protected void assertEqualsGeneratedSourceWithResource(
       final URL expectedResourceUrl, final String className) throws Exception {
-    try {
-      compiler.assertEqualsGeneratedSourceWithResource(expectedResourceUrl, className);
-    } catch (AssertionError error) {
-      System.out.println(compiler.getGeneratedSource(className));
-      throw error;
-    }
+    delegate.assertEqualsGeneratedSourceWithResource(expectedResourceUrl, className);
   }
 
   protected void assertMessage(Message message) {
-    compiler.assertMessage(message);
+    delegate.assertMessage(message);
   }
 
   protected void assertNoError() {
-    boolean match =
-        compiler.getDiagnostics().stream().anyMatch(d -> d.getKind() == Diagnostic.Kind.ERROR);
-    if (match) {
-      fail();
-    }
+    delegate.assertNoError();
   }
 }
